gas 1

+++ src/Gas.sol
@@ -51,9 +51,9 @@ contract GasContract is Ownable, Constants {
     
     struct ImportantStruct {
         uint256 amount;
-        uint256 valueA; // max 3 digits
+        uint32 valueA; // max 3 digits
         uint256 bigValue;
-        uint256 valueB; // max 3 digits
+        uint32 valueB; // max 3 digits
         bool paymentStatus;
         address sender;
     }
